Rename ProjectBridge.ts to ProjectEscrow.ts
