+++ packages/config/src/projects/types/index.ts
@@ -1,8 +1,8 @@
 export * from './News'
 export * from './Project'
-export * from './ProjectBridge'
 export * from './ProjectContracts'
 export * from './ProjectDetails'
+export * from './ProjectEscrow'
 export * from './ProjectLinks'
 export * from './ProjectPermission'
 export * from './ProjectReference'
